+++ sdk/servicebus/service-bus/test/serviceBusClient.spec.ts
@@ -570,7 +570,7 @@ describe("Errors after close()", function(): void {
     );
 
     let errorPeek: string = "";
-    await receiver.peekMessages().catch((err) => {
+    await receiver.peekMessages(1).catch((err) => {
       errorPeek = err.message;
     });
     should.equal(
@@ -605,7 +605,7 @@ describe("Errors after close()", function(): void {
     const sessionReceiver = receiver as SessionReceiver<ReceivedMessageWithLock>;
 
     let errorPeek: string = "";
-    await sessionReceiver.peekMessages().catch((err) => {
+    await sessionReceiver.peekMessages(1).catch((err) => {
       errorPeek = err.message;
     });
     should.equal(
@@ -615,7 +615,7 @@ describe("Errors after close()", function(): void {
     );
 
     let errorPeekBySequence: string = "";
-    await sessionReceiver.peekMessages({ fromSequenceNumber: Long.ZERO }).catch((err) => {
+    await sessionReceiver.peekMessages(1, { fromSequenceNumber: Long.ZERO }).catch((err) => {
       errorPeekBySequence = err.message;
     });
     should.equal(