+++ sdk/storage/storage-file-share/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 12.2.0-preview.1 (2020.07)
+
+- Updated Azure Storage Service API version to 2019-12-12.
+- Support 4 TB files.
+- Added `exists` and `deleteIfExists()` to `ShareClient`, `ShareDirectoryClient`, and `ShareFileClient`.
+- Added `createIfNotExists()` to `ShareClient` and `ShareDirectoryClient`.
+
 ## 12.1.2 (2020.05)
 
 - Fix data corruption failure error [issue #6411](https://github.com/Azure/azure-sdk-for-js/issues/6411) when downloading compressed files. [PR #7993](https://github.com/Azure/azure-sdk-for-js/pull/7993)